test(compiler): isolate MIR verifier regression target
