+++ src/tests/CMakeLists.txt
@@ -67,6 +67,7 @@ if(ZEN_ENABLE_SPEC_TEST)
     target_link_libraries(evmCacheComplexityDemo PRIVATE dtvmcore)
     if(ZEN_ENABLE_MULTIPASS_JIT)
       add_executable(evmJitFrontendTests evm_jit_frontend_tests.cpp)
+      add_executable(evmMirVerifierTests evm_mir_verifier_tests.cpp)
       add_executable(evmRangeAnalyzerTests evm_range_analyzer_tests.cpp)
       add_executable(evmDifferentialTests evm_differential_tests.cpp)
       add_executable(
@@ -113,6 +114,7 @@ if(ZEN_ENABLE_SPEC_TEST)
       target_compile_options(evmCacheTests PRIVATE -fsanitize=address)
       if(ZEN_ENABLE_MULTIPASS_JIT)
         target_compile_options(evmJitFrontendTests PRIVATE -fsanitize=address)
+        target_compile_options(evmMirVerifierTests PRIVATE -fsanitize=address)
         target_compile_options(evmRangeAnalyzerTests PRIVATE -fsanitize=address)
         target_compile_options(evmDifferentialTests PRIVATE -fsanitize=address)
       endif()
@@ -146,6 +148,11 @@ if(ZEN_ENABLE_SPEC_TEST)
         if(ZEN_ENABLE_MULTIPASS_JIT)
           target_link_libraries(
             evmJitFrontendTests
+            PRIVATE compiler dtvmcore gtest_main -fsanitize=address
+            PUBLIC ${GTEST_BOTH_LIBRARIES}
+          )
+          target_link_libraries(
+            evmMirVerifierTests
             PRIVATE compiler dtvmcore gtest_main -fsanitize=address
             PUBLIC ${GTEST_BOTH_LIBRARIES}
           )
@@ -226,6 +233,12 @@ if(ZEN_ENABLE_SPEC_TEST)
         if(ZEN_ENABLE_MULTIPASS_JIT)
           target_link_libraries(
             evmJitFrontendTests
+            PRIVATE compiler dtvmcore gtest_main -fsanitize=address
+                    -static-libasan
+            PUBLIC ${GTEST_BOTH_LIBRARIES}
+          )
+          target_link_libraries(
+            evmMirVerifierTests
             PRIVATE compiler dtvmcore gtest_main -fsanitize=address
                     -static-libasan
             PUBLIC ${GTEST_BOTH_LIBRARIES}
@@ -320,6 +333,11 @@ if(ZEN_ENABLE_SPEC_TEST)
           PUBLIC ${GTEST_BOTH_LIBRARIES}
         )
         target_link_libraries(
+          evmMirVerifierTests
+          PRIVATE compiler dtvmcore gtest_main
+          PUBLIC ${GTEST_BOTH_LIBRARIES}
+        )
+        target_link_libraries(
           evmRangeAnalyzerTests
           PRIVATE compiler dtvmcore gtest_main
           PUBLIC ${GTEST_BOTH_LIBRARIES}
@@ -381,6 +399,7 @@ if(ZEN_ENABLE_SPEC_TEST)
     add_test(NAME evmCacheTests COMMAND evmCacheTests)
     if(ZEN_ENABLE_MULTIPASS_JIT)
       add_test(NAME evmJitFrontendTests COMMAND evmJitFrontendTests)
+      add_test(NAME evmMirVerifierTests COMMAND evmMirVerifierTests)
       add_test(NAME evmRangeAnalyzerTests COMMAND evmRangeAnalyzerTests)
       add_test(NAME evmDifferentialTests COMMAND evmDifferentialTests)
     endif()
